More Dir::get test work.
The diamond test case is tricky, because it's not simple to construct a diamond-shaped dependency. I might want to manually build the Dir tree in the test case (but I want to use DirBuild to make the files).

I'm also unsure how it will work when a directory is materialized RO into the diagnostics dir. It needs to be written into Freezer -- does that need to be a method on Freezer? Trying to avoid a Freezer <-> Dir <-> DiagnosticDir ownership  loop.